backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
supabase==2.0.3
slowapi==0.1.9
litellm==1.17.0
openai==1.3.0
numpy==1.26.2
pydantic-settings==2.1.0
structlog==23.2.0

